Feat: Workspace policy

+++ app/Http/Controllers/WorkspaceController.php
@@ -13,6 +13,7 @@ use App\Models\Workspace;
 use App\Traits\HasFile;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use Inertia\Response;
 
 class WorkspaceController extends Controller
@@ -82,6 +83,7 @@ class WorkspaceController extends Controller
 
     public function edit(Workspace $workspace): Response
     {
+        Gate::authorize('update_workspace', $workspace);
         return Inertia(
             'Workspaces/Setting',
             [
@@ -99,6 +101,7 @@ class WorkspaceController extends Controller
 
     public function update(Workspace $workspace, WorkspaceRequest $request): RedirectResponse
     {
+        Gate::authorize('update_workspace', $workspace);
         $workspace->update([
             'name' => ($name = $request->name),
             'slug' => str()->slug($name . str()->uuid(10)),
@@ -114,6 +117,7 @@ class WorkspaceController extends Controller
 
     public function destroy(Workspace $workspace): RedirectResponse
     {
+        Gate::authorize('delete_workspace', $workspace);
         $this->delete_file($workspace, 'cover');
         $this->delete_file($workspace, 'logo');
 
@@ -127,6 +131,7 @@ class WorkspaceController extends Controller
 
     public function member_store(Workspace $workspace, Request $request): RedirectResponse
     {
+        Gate::authorize('member_workspace', $workspace);
         $request->validate([
             'email' => ['required', 'email', 'string'],
         ]);
